Revert "Merge branch 'patch_trajectory' into merge_diffusion1d.DiffusionProcess"
Reason for change in the first place
---------------------------------------
For 1D processes, attribute method diffusion() returns a 'float' that does not have attribute
'dtype'.

Reason for revert
-------------------
this originally caused an Exception when testing method trajectory with 1D processes, but at the
time of this commit attribute method trajectory is only tested against multidimensionnal processes. Additionally, this change introduces more exceptions for tests for processes with d>1.
This issue should be delt with in a separate PR.

This reverts commit 6f31acbe5be31a9e6bdf1f3c2be670797b2c0e00, reversing
changes made to 8235252fbe32d3797d6e5a654cbca4ecbff3c38d.